Do not re-typecheck closures in the presence of dangling type variables
It is never safe to spawn a new constraint system in the presence of
free type variables.  In this case, if CSDiag re-typechecked a closure
involving any type variables in a bound generic type, the type checker
would skip opening those type variables into the constraint system on
the second go-around but would still create constraints referencing the
variables as though they were a part of its domain.

In particular, sanitize casts before attempting to get possible
solutions by filling in ErrorType and teach the RAII expression cleaners
to clean up after casts.