_StringGuts: Introduce _StringObject encapsulation.
Add a _StringObject struct to wrap the BridgeObject and encapsulate
all the low level, bit twiddly, and unsafe operations. This is meant
to be a "transparent" structure, that is all methods and properties
should always be inlined and directly emitted. It just encapsulates
operations that are otherwise tricky to implement correctly.