WIP IRGen: preserve calling convention when importing functions
We do not currently preserve the calling convention when importing a
foreign function.  This is problematic as a call made with a mismatched
calling convention is UB.  The optimizer in LLVM then is permitted to
optimize that UB into a call elision.  For platforms where the default
calling convention is non-C (e.g. Linux ARMHF/Windows ARM default to ARM
AAPCS VFP CC rather than C) this will result in invalid function calls
outside of swift.