[ClangImporter] Translate Clang's enum_extensibility to exhaustivity
If a CF_ENUM has no enum_extensibility annotation, default to
exhaustive in Swift 3 and 4 and non-exhaustive in Swift 5. Once Swift
5 comes out properly, unannotated CF_ENUMs will be assumed to be
non-exhaustive everywhere, generating warnings in Swift 4 code and
errors in Swift 5.

There's one extra wrinkle in this implementation: whether or not the
annotations themselves are printed in the generated interface depends
on whether the current mode is Swift 4 or Swift 5. Only the
non-default value is printed.

Still to do: test and fix up the use of multiple enum_extensibility
annotations, possibly with API notes. This is interesting if the
Foundation folks decide to bake enum_extensibility(open) into NS_ENUM
in the future; there should be an out people can use to declare
exhaustive enums in C that's backwards-compatible.