[runtime/abi] Add tagged BridgeObject recognition to 32-bit.
BridgeObject is useful as a builtin to distinguish objects from values
(i.e. does retain/release need to track lifetime) as well as expose an
additional bit to denote ObjC vs Swift tracking. It just so happens
that there are no tagged pointers on 32-bit, so BridgeObject didn't
historically provide a tagged pattern on 32-bit. This adds that
ability, as it will be useful for String in Swift 5 as well as any
other structures that need a built-in way of distinguishing values
from references as well as ObjC vs Swift management.