IUO: Create disjunctions for optional choices.
When binding an optional value, or function that returns an optional
value, if that value was produced from a decl that was declared an
IUO, create a disjunction.

This change appears to regress a couple diagnostics tests. I've opened
a bug report (https://bugs.swift.org/browse/SR-6626) to investigate
these so that we can produce the previous diagnostic again at some
point.