IUO: Do not short-circuit disjunctions if we had to apply fixes.
There are cases where the disjunctions created for IUOs can be solved
with the Optional choice if a fix is applied. We do not what to stop
with this solution as we may be able to solve the constraint system
without a fix by selecting the non-Optional side.

This restores a couple diagnostics that had changed in a previous
commit, but also slightly regresses another.