Initial checkin of the TFCanonicalizeCFG pass.
This introduces the (still in early development) pass that canonicalizes the
CFG of a SIL function into a well structured single-entry-single-exit regions.
The SILPass that wraps it won't be immediately used by the compiler itself, but
is used to allow us to write testcases.  An initial test is included but more
are necessary and will be added over time.

PiperOrigin-RevId: 179817076