Initial checkin of TFLowerGraph pass, which transforms a lowered SIL function into a tensorflow graph function.  It is still very early (it only supports single basic block functions).
Also, there are no tests for this, I'm not sure of a good way to write tests,
given the output is a binary protobuf.  Once we figure that out though it can
be added in follow-on patches.

update: fixed an accidental removal of the SWIFT_ENABLE_TENSORFLOW comment in
DiagnosticsSIL.def.
PiperOrigin-RevId: 179852046