IUO: Generate Optional<T> rather than ImplicitlyUnwrappedOptional<T>.
Stop creating ImplicitlyUnwrappedOptional<T> so that we can remove it
from the type system.

Enable the code that generates disjunctions for Optional<T> and
rewrites expressions based on the original declared type being 'T!'.

This effectively means ImplicitlyUnwrappedOptional<T> is no longer
part of the type system, although I haven't actually removed all of
the code dealing with it yet.

ImplicitlyUnwrappedOptionalType is is dead, long live implicitly
unwrapped Optionals!

Resolves rdar://problem/33272674.