IUO: Force cases where we go from optional to non-optional.
This used to look at the type and force only IUO cases and then hit an
unreachable on plain optionals.

We should never reach this point for plain optional declarations,
though, so we should just be able to force value conversion from
optional-to-non-optional, knowing that Sema only allowed those things
declared as implicitly unwrapped optionals though.