Made CompilerRunTime.swift depend on the CTensorFlow module, which exports the TF C API for CompilerRunTime.swift to call (e.g. TF_NewStatus()).
This is originally hongm@'s CL; I patched it in and tinkered with the necessary toolchain changes to get it mostly working with our current build infrastructure.

This CL has two major-ish changes to the toolchain:

- It adds the -Xcc option to sil-opt so that we can pass additional paths to ClangImporter when using that testing tool. We'll want to upstream this change soon as part of the "changes needed for content-addressable storage" that I'm collecting.

- It modifies the lit runner script so that it *almost* works with TensorFlow runtime tests. Importing TensorFlow successfully finds the headers for the underlying CTensorFlow module; the only missing piece is the library for swiftc/clang to link to. That's a slightly harder problem, because we don't have a single cc_library target whose .a file we can snag; we have to grab the transitive objects from a handful of targets and put them into a complete library that looks more or less like the one that someone using open-source TensorFlow C bindings would see.

I'm punting on that last part for now because what's most important here is unblocking the tf-compiler team and ensuring we have *some* test coverage, which the current swift_linux_test targets do achieve even if we're not using the lit runner.

PiperOrigin-RevId: 180557841