SwiftSyntax: make Syntax node Hashable.
This gives Syntax node a unique identifier and makes them friendly
to Set and Dictionary.

We used to compare RawSyntax's references to decide if two Syntax nodes
are identical. This is insufficient because (1) a RawSyntax can be
shared among several Syntax nodes, at least in theory; and (2) two
instances of RawSyntax can still have the same content.