[AST] Perf: Improve getDesugaredType() efficiency
Make getDesugaredType() as fast as possible (short of a major redesign).
With the old way:

1) Switching over the sugared types turned into a frequently
mispredicted branch because the sugar in the type system essentially
random as far as the processor is concerned.
2) Storing the underlying/singlely-desugared type at different offsets
in memory adds more code bloat and misprediction.

Short of a major redesign to avoid pointer chasing, or inlining the fast
path of getDesugaredType(), this is about as fast as the method will get.