[ClangImporter] Stop copying EnumInfo strings into Swift's ASTContext
They're all kept alive by Clang's ASTContext already. Since the enum
info is only accessible through the Clang importer, that's a
sufficient lifetime. (Note that this only works because all of the
strings are derived from text actually present in the C sources:
either attribute arguments or names of declarations.)

Fixes a small TODO and probably saves a tiny bit of memory usage.