Add a new "swift5-transition" checkout config.
This combination of branches uses the master-llvm-swift5-transition branch
of Swift in combination with the swift-5.0-branch from the LLVM repos.
This will be used on a temporary basis to test Swift with the new LLVM
branches and can be removed when we switch over to develop against the
new LLVM branches by default.