Remove IUO nested inside other type from benchmark.
SE-0054 disallowed nesting IUO types within other types. We currently
give a warning for this, but eventually we'll either emit an error or
interpret the '!' as '?' (a plain optional).

Either way, this particular benchmark doesn't test anything different
than the one that uses plain optionals, so we should be able to remove it.