SILGen: Fix crash when emitting implicit 'super.init()' delegation to throwing or failing initializer
If the base class defines a no-argument initializer, we allow the
'super.init()' call to be omitted. However, SILGen would generate
invalid SIL if the base class initializer was throwing or failing.

The problem was that we would attempt to insert instructions
after a try_apply or cond_br, which is invalid, because we would
incorrectly save and restore the insertion point when emitting the
delegation. This is incorrect, because a delegation to a throwing
or failing initializer emits new basic blocks.

Fixes <https://bugs.swift.org/browse/SR-5974>,
<rdar://problem/36256335>.