Make TypeChecker::coerceToRValue walk into the subexpr of ForceValueExpr.
This results in the load for the coercion ending up below the force.

I don't have a case where this makes a difference on master, with the
new IUO implementation this gets us closer to generating the same code
that we generate on master, where we rely on IUO types to force
results after we've coerced lvalues to rvalues.