Add support for specifying driver or frontend specific arguments via environment options.
Previously, we only had SWIFT_TEST_OPTIONS for this purpose and it applied
whether we were calling the driver or the frontend. This doesn't work well for
frontend arguments since in 1 context you pass in $ARGUMENT and in the other
context you have to pass in -Xfrontend $ARGUMENT. Instead now there are two
additional env vars that allows one to work around these issues:

SWIFT_DRIVER_TEST_OPTIONS
SWIFT_FRONTEND_TEST_OPTIONS

These only pass their arguments to the proper locations. *NOTE* I did not get
rid of SWIFT_TEST_OPTIONS since I did not want to stomp on any tools or
functionality that relies upon it and it seems like potentially it could be
useful (consider -Xllvm options which pass the same to the driver and the
frontend).

rdar://34222540